% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{excalidraw_consent}
\alias{excalidraw_consent}
\alias{excalidraw_unconsent}
\title{Consent to using shinidraw}
\usage{
excalidraw_consent(force = FALSE)

excalidraw_unconsent()
}
\arguments{
\item{force}{Force consent}
}
\value{
TRUE if consent is given, FALSE otherwise
}
\description{
Consent to using shinidraw
}
\examples{
if (interactive()) {
  excalidraw_consent()
  excalidraw_unconsent()
}
}
