% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excal_existing_project.R
\name{excal_existing_project}
\alias{excal_existing_project}
\title{List existing project}
\usage{
excal_existing_project()
}
\description{
List existing project saved in the shinidraw project folder
}
\examples{
if (interactive()) {
  excal_new_or_existing()
}
}
