% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{excalidraw_new_project}
\alias{excalidraw_new_project}
\alias{excalidraw_open_project}
\alias{excalidraw_open_file}
\alias{excalidraw}
\title{Run and save excalidraw projects}
\usage{
excalidraw_new_project(
  project,
  auto_save = TRUE,
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  talkative = FALSE
)

excalidraw_open_project(
  project,
  auto_save = TRUE,
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  talkative = FALSE
)

excalidraw_open_file(
  excalidraw,
  auto_save = TRUE,
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  talkative = FALSE
)

excalidraw(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  talkative = FALSE
)
}
\arguments{
\item{project}{name to the project to load or create.}

\item{auto_save}{whether to save the project automatically.}

\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{talkative}{whether to print messages when saving.}

\item{excalidraw}{path to the excalidraw file to load.}
}
\description{
Run and save excalidraw projects
}
