% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excalidraw_existing_project.R
\name{excalidraw_existing_project}
\alias{excalidraw_existing_project}
\title{List existing project}
\usage{
excalidraw_existing_project()
}
\description{
List existing project saved in the shinidraw project folder
}
\examples{
if (interactive()) {
  excalidraw_new_or_existing()
}
}
