% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excalidraw_existing_project.R
\name{excalidraw_get_project}
\alias{excalidraw_get_project}
\title{Get the path to the Excalidraw file}
\usage{
excalidraw_get_project(project_name)
}
\arguments{
\item{project_name}{The name of the project}
}
\description{
Get the link to the excalidraw file
}
