% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_examples.R
\name{excalidraw_helloworld}
\alias{excalidraw_helloworld}
\title{examples}
\usage{
excalidraw_helloworld()
}
\value{
The return value, if any, from executing the utility.
}
\description{
A utils function
}
