% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_excalidraw.R
\name{mod_excalidraw_ui}
\alias{mod_excalidraw_ui}
\alias{mod_excalidraw_server}
\title{main UI Function}
\usage{
mod_excalidraw_ui(id, path_to_excalidraw, with_auto_save = TRUE)

mod_excalidraw_server(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{path_to_excalidraw}{Path to the excalidraw file to load.}

\item{with_auto_save}{Whether to save the project automatically.}
}
\description{
A shiny Module.
}
