% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{consent}
\alias{consent}
\alias{unconsent}
\title{Consent to using shinidraw}
\usage{
consent(force = FALSE)

unconsent()
}
\arguments{
\item{force}{Force consent}
}
\value{
TRUE if consent is given, FALSE otherwise
}
\description{
Consent to using shinidraw
}
\examples{
if (interactive()) {
	consent()
	unconsent()
}
}
